+++ src/main/java/net/frozenblock/wilderwild/tag/WWBlockTags.java
@@ -28,6 +28,15 @@ public final class WWBlockTags {
 	public static final TagKey<Block> SOUND_FLOWER = bind("sound/flower");
 	public static final TagKey<Block> SOUND_LEAVES = bind("sound/leaves");
 	public static final TagKey<Block> SOUND_SAPLING = bind("sound/sapling");
+	public static final TagKey<Block> SOUND_CACTUS = bind("sound/cactus");
+	public static final TagKey<Block> SOUND_COARSE_DIRT = bind("sound/coarse_dirt");
+	public static final TagKey<Block> SOUND_ICE = bind("sound/ice");
+	public static final TagKey<Block> SOUND_FROSTED_ICE = bind("sound/frosted_ice");
+	public static final TagKey<Block> SOUND_MUSHROOM = bind("sound/mushroom");
+	public static final TagKey<Block> SOUND_MUSHROOM_BLOCK = bind("sound/mushroom_block");
+	public static final TagKey<Block> SOUND_SANDSTONE = bind("sound/sandstone");
+	public static final TagKey<Block> SOUND_LILY_PAD = bind("sound/lily_pad");
+	public static final TagKey<Block> SOUND_MELON = bind("sound/melon");
 
 	public static final TagKey<Block> HOLLOWED_LOGS = bind("hollowed_logs");
 	public static final TagKey<Block> HOLLOWED_LOGS_DONT_BURN = bind("hollowed_logs_dont_burn");